refactor!: changing block height type from uint64 to int64 (#979)
* changing height from uint64 to int64

* remove unnecessay type castings

+++ examples/polkadot/push_wasm_client_code_test.go
@@ -22,7 +22,7 @@ import (
 )
 
 const (
-	heightDelta      = uint64(20)
+	heightDelta      = 20
 	votingPeriod     = "30s"
 	maxDepositPeriod = "10s"
 )
